The size of the wispering memory is taken from the configuration now

+++ io-mokamint-node-messages/src/main/java/io/mokamint/node/messages/MessageMemories.java
@@ -32,7 +32,7 @@ public abstract class MessageMemories {
 	 * @return the message memory
 	 * @throws IllegalArgumentException if {@code size} is negative
 	 */
-	public static MessageMemory of(int size) {
+	public static MessageMemory of(long size) {
 		return new MessageMemoryImpl(size);
 	}
 }
